pyproject.toml
[project]
name = "djangorenteasy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref",
    "certifi",
    "charset-normalizer",
    "crispy-tailwind",
    "dj-database-url",
    "django>=6.0.2",
    "django-admin-honeypot-updated-2021",
    "django-browser-reload",
    "django-cloudinary-storage",
    "django-crispy-forms",
    "django-debug-toolbar",
    "django-jazzmin",
    "gunicorn",
    "idna",
    "pillow",
    "psycopg2-binary",
    "python-dotenv",
    "requests",
    "sqlparse",
    "urllib3",
    "whitenoise",
]
